RemoteControlGUI/requirements.txt
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
pyserial==3.5
